requirements.txt
fastapi
uvicorn[standard]

supabase

python-jose[cryptography]
passlib[bcrypt]

pydantic
python-dotenv

email-validator